% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-RFLP.R
\name{add_RFLP}
\alias{add_RFLP}
\title{Add RFLP enzymes to iSTOP data}
\usage{
add_RFLP(iSTOP, recognizes = "c", width = 150, enzymes = NULL, cores = 1)
}
\arguments{
\item{iSTOP}{A dataframe resulting from \link{locate_PAM}.}

\item{recognizes}{The base recognized by the enzyme. One of "c" or "t".
In terms of base editing, "t" will return enzymes that only cut after editing,
whereas "c" will return enzymes that only cut before editing.}

\item{width}{An integer specifiying range of unique cutting.}

\item{enzymes}{A dataframe of enzymes considered. Defaults to NULL which will
use an enzyme dataset included in the iSTOP package.}

\item{cores}{Number of cores to use with \link[=pblapply]{pbapply}}
}
\description{
Add RFLP enzymes to iSTOP data
}
\details{
This function only works if changed base can be distinguished. Current strategy
is to mark the changed base by converting it to lower case, and constructing
an \code{enzymes} table that has each corresponding base chaned to lowercase.
See \code{enzymes} definition below for more information.
}
