% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS-coordinates.R
\name{CDS}
\alias{CDS}
\alias{CDS_Athaliana_BioMart_plantsmart28}
\alias{CDS_Celegans_UCSC_ce11}
\alias{CDS_Dmelanogaster_UCSC_dm6}
\alias{CDS_Drerio_UCSC_danRer10}
\alias{CDS_Hsapiens_UCSC_hg38}
\alias{CDS_Mmusculus_UCSC_mm10}
\alias{CDS_Rnorvegicus_UCSC_rn6}
\alias{CDS_Scerevisiae_UCSC_sacCer3}
\title{CDS coordinates}
\usage{
CDS(txdb, orgdb, gene_id = "ENTREZID", gene = "SYMBOL", fix_chr = NULL,
  fix_ids = NULL)

CDS_Athaliana_BioMart_plantsmart28()

CDS_Celegans_UCSC_ce11()

CDS_Dmelanogaster_UCSC_dm6()

CDS_Drerio_UCSC_danRer10()

CDS_Hsapiens_UCSC_hg38()

CDS_Mmusculus_UCSC_mm10()

CDS_Rnorvegicus_UCSC_rn6()

CDS_Scerevisiae_UCSC_sacCer3()
}
\arguments{
\item{txdb}{\link[GenomicFeatures:TxDb]{TxDb}
Transcript database object.}

\item{orgdb}{\link[AnnotationDbi:AnnotationDb]{OrgDb}
Organism annotation database object.}

\item{gene_id}{\link[assertthat:is.string]{String}
Name of gene ID column in \code{orgdb}. See \link[AnnotationDbi:columns]{columns}.}

\item{gene}{\link[assertthat:is.string]{String}
Name of gene symbol column in \code{orgdb}. See \link[AnnotationDbi:columns]{columns}}

\item{fix_chr}{A function to fix the \code{chr} column.
This function should accept a vector and return a
vector of the same length. See 'Details'.}

\item{fix_ids}{A function to fix the gene_id column.
This function should accept a vector and return a
vector of the same length. See 'Details'.}
}
\value{
A data.frame with the following columns where each row represents
a single exon:
\itemize{
\item \strong{COLUMN-NAME} \emph{\code{DATA-TYPE}} DESCRIPTION
\item \strong{gene}    \emph{\code{chr}} Gene symbol
\item \strong{tx}      \emph{\code{chr}} Transcript symbol
\item \strong{exon}    \emph{\code{int}} Exon rank in gene
\item \strong{ncds}    \emph{\code{int}} Exon rank in CDS
\item \strong{gene_id} \emph{\code{chr}} Gene identifier
\item \strong{tx_id}   \emph{\code{int}} Transcript identifier
\item \strong{exon_id} \emph{\code{int}} Exon identifier
\item \strong{cds_id}  \emph{\code{int}} CDS identifier
\item \strong{chr}     \emph{\code{chr}} Chromosome
\item \strong{strand}  \emph{\code{chr}} Strand (+/-)
\item \strong{start}   \emph{\code{int}} CDS coordinate start (always < \strong{end})
\item \strong{end}     \emph{\code{int}} CDS coordinate end (always > \strong{start})
}
}
\description{
Compiles a data frame of coding sequence coordinates for all exons in a
\link[GenomicFeatures:TxDb]{TxDb} object. Common gene names are also included
from an \link[AnnotationDbi:AnnotationDb]{OrgDb} object. Note that CDS coordinates are
nearly equivalent to exon coordinates except the first and last exons that
contain coding sequence should have coordinates that begin with the
translation start codon and end with the stop codon.
}
\details{
If constructing a CDS data.frame manually (1) each row should
represent the coordinates of a single exon, (2) exons should be numbered in
order with reference to the transcript's strand (i.e. the first exon should
include the start codon), (3) and the first and last exon coordinates should
beging with the start codon and end with the stop codon.

Pre-defined CDS data.frame builders are provided. They
are named \code{CDS_<Species>_<data-source>_<genome-assembly-ID>()}

The arguments \code{fix_chr} and \code{fix_ids} are provided for convenience
to fix chromosome names or gene IDs in the TxDb to match those in the
OrgDb. The fix functions are applied as a transformation of the TxDb \code{chr} and
\code{gene_id} columns. For example, this functionality is used
in \code{CDS_Athaliana_BioMart_plantsmart28()} to provide a mapping of chromosome
names from the TxDb format to the OrgDb format (1 = Chr1, Mt = ChrM ...). The
\code{fix_ids} functionality is used in \code{CDS_Dmelanogaster_UCSC_dm6()} to remove a
trailing '.1' that was added to every ID in the TxDb.
}
