% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-genomic-sequence.R
\name{get_genomic_sequence}
\alias{get_genomic_sequence}
\title{Get Genomic Sequence}
\usage{
get_genomic_sequence(at, add_5prime, add_3prime, genome, chr, strand)
}
\arguments{
\item{at}{<int:vector> A coordinate in a chromosome. First base is at 1.}

\item{add_5prime}{<int:vector> Number of bases to add to 5' side of `at`
with respect to `strand`.}

\item{add_3prime}{<int:vector> Number of bases to add to 3' side of `at`
with respect to `strand`.}

\item{genome}{<BSgenome>   A BSgenome object of sequence data.}

\item{chr}{<chr:vector> Chromosome names. Must match names in `BSgenome`.}

\item{strand}{<chr:vector> Strand orientation. All strand values must
be '+' or '-'.}
}
\description{
From a [BSgenome::BSgenome] object access upstream and downstream sequence
with respect to strand at a chromosomal coordinate. A '-' character will be
added for each position outside of sequence boundary.
}
