% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate-codons.R
\name{locate_codons}
\alias{locate_codons}
\title{Locate codons in transcripts}
\usage{
locate_codons(
  cds,
  genome,
  codons = c("CAA", "CAG", "CGA", "TGG", "TGG"),
  positions = c(1L, 1L, 1L, 2L, 3L),
  switch_strand = c(F, F, F, T, T),
  cores = getOption("mc.cores", 1L)
)
}
\arguments{
\item{cds}{A data.frame of CDS coordinates. See \link[=CDS#Value]{CDS} for details
on required columns.}

\item{genome}{A \link[BSgenome:BSgenome-class]{BSgenome} sequence database,
or a \link[=readDNAStringSet]{Biostrings}. CDS coordinates in \code{cds} should
correspond to this genome assembly.}

\item{codons}{A character vecor of codons to consider. Defaults to
\code{c('CAA', 'CAG', 'CGA', 'TGG', 'TGG')} which are all of the codons
with iSTOP targetable bases.}

\item{positions}{An integer vector of positions in \code{codons} to use for
returned coordinates. Should be the same length as \code{codons}. Defaults to
\code{c(1L, 1L, 1L, 2L, 3L)} which are the positions in \code{codons} that can be
targeted with iSTOP.}

\item{switch_strand}{A logical vector indicating whether the corresponding
\code{codon} is targeted on the opposite strand. Determines the resulting value of
\code{sg_strand}. Should be the same length as \code{codons}. Defaults to \code{c(F, F, F, T, T)}}

\item{cores}{Number of cores to use for parallel processing with \link[pbapply:pbapply]{pblapply}}
}
\value{
A data.frame with the following columns where each row represents
a single targetable coordinate (codons with multiple targetable coordinates
will have a separate row for each):
\itemize{
\item \strong{COLUMN-NAME} \emph{\code{DATA-TYPE}} DESCRIPTION
\item \strong{tx}           \emph{\code{chr}} Transcript symbol
\item \strong{gene}         \emph{\code{chr}} Gene symbol
\item \strong{exon}         \emph{\code{int}} Exon rank in gene
\item \strong{pep_length}   \emph{\code{int}} Lenth of peptide
\item \strong{cds_length}   \emph{\code{int}} Length of CDS DNA
\item \strong{chr}          \emph{\code{chr}} Chromosome
\item \strong{strand}       \emph{\code{chr}} Strand (+/-)
\item \strong{sg_strand}    \emph{\code{chr}} Guide strand (+/-)
\item \strong{aa_target}    \emph{\code{chr}} One letter code of targeted amino acid
\item \strong{codon}        \emph{\code{chr}} Three letter codon
\item \strong{aa_coord}     \emph{\code{int}} Coordinate of targeted amino acid (start == 1)
\item \strong{cds_coord}    \emph{\code{int}} Coordinate of targeted base in CDS (start == 1)
\item \strong{genome_coord} \emph{\code{int}} Coordinate of targeted base in genome
\item \strong{NMD_pred}     \emph{\code{lgl}} Is nonsense-mediated-decay predicted (i.e. target base is >56 bases upstream of last exon junction)
\item \strong{rel_position} \emph{\code{dbl}} Relative position in CDS of the targeted base
}
}
\description{
Locate genome coordinates of a desired set of codons given a data.frame of
CDS coordinates (see \link{CDS}), and its corresponding
\link[BSgenome:BSgenome-class]{BSgenome} sequence database.
}
\details{
Each transcript is processed independently based on the \code{tx}
column of the \code{cds} data.frame.

\strong{CDS validation} - Each CDS is validated by checking that the coordinates
result in a DNA sequence that begins with a start codon, ends with a stop codon,
has a length that is a multiple of 3 and has no internal in-frame stop codons.

\strong{No match?} - If a CDS passes validation, but does not have any of the
considered codons, the transcript will still be included in the resulting
data.frame, but coordinates will be missing (i.e. \code{NA}).
}
