% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate-PAM.R
\name{locate_PAM}
\alias{locate_PAM}
\title{Locate PAM near a genomic coordinate}
\usage{
locate_PAM(codons, genome, spacing = PAM_spacing(),
  PAM = PAM_patterns_default(), flanking = 150, keep_PAM = FALSE)
}
\arguments{
\item{codons}{A dataframe resulting from \link{locate_codons}.}

\item{genome}{A \link[BSgenome:BSgenome]{BSgenome} sequence database,
or a \link[=readDNAStringSet]{Biostrings}. Used to extract the genomic sequence context
for each genomic coordinate in \code{codons}.}

\item{spacing}{The result of \link{PAM_spacing}, which is a named list of two-element
numeric vectors. Each two-element vector defines a range of spacing between an
edited base and the PAM (i.e. the minimum and maximum allowable nucleotides between
the targeted base and the PAM). Each pair of ranges is checked in order with the
first range prioritized over the next. Defaults to \code{PAM_spacing()}.}

\item{PAM}{A named list of PAM patterns to be considered (see \link{PAM_pattern}). Names
correspond to resulting column names in the returned dataframe. Default PAM patterns
include \code{.GG}, \code{.GA}, \code{.GCG}, \code{.GAG}, \code{..G[AG][AG]T}, and \code{...[AG][AG]T}.}

\item{flanking}{An number specifying how much flanking genomic context to return
in the resulting dataframe. Defaults to 150.}

\item{keep_PAM}{Logical. Should the PAM sequence be included in the guides?
Defaults to \code{FALSE}.}
}
\description{
Uses the results of \link{locate_codons} to determine whether there is an available
PAM with desired spacing from the genomic coordinate of the targeted base.
}
