% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate-PAM.R
\name{PAM_pattern}
\alias{PAM_pattern}
\alias{PAM_patterns_default}
\title{PAM patterns}
\usage{
PAM_pattern(pattern, width)

PAM_patterns_default(
  sgNGG = PAM_pattern(".GG", 3),
  sgNGA = PAM_pattern(".GA", 3),
  sgNGCG = PAM_pattern(".GCG", 4),
  sgNGAG = PAM_pattern(".GAG", 4),
  sgNNGRRT = PAM_pattern("..G[AG][AG]T", 6),
  sgNNNRRT = PAM_pattern("...[AG][AG]T", 6),
  ...
)
}
\arguments{
\item{pattern}{A regular expression used to identify a PAM. See the \link[stringi:about_search_regex]{regex documentation}
for extensive help on regular expressions.}

\item{width}{A single number indicating the expected width of the PAM sequence.}

\item{sgNGG}{Pattern for NGG. Defaults to \code{PAM_pattern(".GG", 3)}}

\item{sgNGA}{Pattern for NGA. Defaults to \code{PAM_pattern(".GA", 3)}}

\item{sgNGCG}{Pattern for NGCG. Defaults to \code{PAM_pattern(".GCG", 4)}}

\item{sgNGAG}{Pattern for NGAG. Defaults to \code{PAM_pattern('.GAG', 4)}}

\item{sgNNGRRT}{Pattern for NNGRRT. Defaults to \code{PAM_pattern('..G[AG][AG]T', 6)}}

\item{sgNNNRRT}{Pattern for NNNRRT. Defaults to \code{PAM_pattern('...[AG][AG]T', 6)}}

\item{...}{Additional PAM patterns to include in \code{PAM_patterns_default()}}
}
\description{
Specify PAM patterns for use with \link{locate_PAM}
}
