% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate-PAM.R
\name{PAM_spacing}
\alias{PAM_spacing}
\title{PAM spacing specification}
\usage{
PAM_spacing(optimal = c(14, 14), good = c(13, 15), ok = c(12, 16), ...)
}
\arguments{
\item{optimal}{defaults to \code{c(14, 14)}}

\item{good}{defaults to \code{c(13, 15)}}

\item{ok}{defaults to \code{c(12, 16)}}

\item{...}{Additional custom named ranges can be included.}
}
\description{
Specify preferred PAM spacing when using \link{locate_PAM}.
}
\details{
\link{locate_PAM} uses these spacing ranges to optimize PAM selection when designing
guide sequences. The priority for design is \verb{optimal > good > ok > ...}. All arguments should
be two-element numeric vectors indicating a minimum and maximum of the range. The default
settings gives priority to PAMs that place the target base in the middle of the five base
range of high efficiency BE3-Cas9 editing.

Note that a spacing of 14 means 14 nucleotides between the target base the first base in the PAM.
For example, \code{cNNNNNNNNNNNNNN.GG} has a spacing of 14 between the targeted \code{c} and the \code{.GG} PAM
pattern.
}
