% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-off-target.R
\name{search_off_target}
\alias{search_off_target}
\title{Search for candidate off target sequences.}
\usage{
search_off_target(
  guides,
  genome,
  chromosomes,
  fixed_start = 9,
  fixed_end = 20,
  max_mismatch = 2,
  secondary_filter = NULL,
  cores = 1
)
}
\arguments{
\item{guides}{Character vector of sequences to search for fuzzy matches in the genome.}

\item{genome}{A BSgenome}

\item{chromosomes}{Character vector of chromosome names to consider when searching the genome.}

\item{fixed_start}{Beginning of fixed band (no tolerated mismatches). Larger band reduces search time.}

\item{fixed_end}{End of fixed band (no tolerated mismatches). Larger band reduces search time.}

\item{max_mismatch}{Number of tolerated mismatches outside of the fixed band.}

\item{secondary_filter}{A regular expression passed to \link[stringr:str_detect]{stringr::str_detect} used to filter for
off-targets that match this pattern (e.g. 'GG$' will require the matching sequence to end with 'GG').}
}
\description{
Search for candidate off target sequences allowing for a fixed-band to reduce computation time
and a set number of allowed mismatches.
}
