% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-RFLP.R
\name{add_RFLP}
\alias{add_RFLP}
\title{Add RFLP enzymes to iSTOP data}
\usage{
add_RFLP(iSTOP, width = 150, enzymes = NULL, cores = 1)
}
\description{
Add RFLP enzymes to iSTOP data
}
\details{
This function only works if changed base can be distinguished. Current strategy
is to mark the changed base by converting it to lower case, and constructing
an \code{enzymes} table that has each corresponding base chaned to lowercase.
See \code{enzymes} definition below for more information.
}
