% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS-coordinates.R
\name{CDS}
\alias{CDS}
\alias{CDS_example}
\alias{CDS_Celegans_UCSC_ce11}
\alias{CDS_Dmelanogaster_UCSC_dm6}
\alias{CDS_Drerio_UCSC_danRer10}
\alias{CDS_Hsapiens_UCSC_hg38}
\alias{CDS_Mmusculus_UCSC_mm10}
\alias{CDS_Rnorvegicus_UCSC_rn6}
\alias{CDS_Scerevisiae_UCSC_sacCer3}
\alias{CDS_Athaliana_BioMart_plantsmart28}
\title{Generate a CDS coordinates table}
\usage{
CDS(tx, gene, tx_cols, gene_cols, shift_start = 1L, shift_end = 0L)

CDS_example()

CDS_Celegans_UCSC_ce11()

CDS_Dmelanogaster_UCSC_dm6()

CDS_Drerio_UCSC_danRer10()

CDS_Hsapiens_UCSC_hg38()

CDS_Mmusculus_UCSC_mm10()

CDS_Rnorvegicus_UCSC_rn6()

CDS_Scerevisiae_UCSC_sacCer3()

CDS_Athaliana_BioMart_plantsmart28()
}
\arguments{
\item{tx}{A URL to a genome's transcript reference file. This table must have
tab separated fields and contain, identifiers for each transcript,
chromosome, strand, CDS start/end, and exon start/end information.
There should be only one row per transcript and exon start/end columns should
contain comma separated cooordinates for each exon in the transcript.
An example file can be found
\href{http://hgdownload.soe.ucsc.edu/goldenPath/hg38/database/knownGene.txt.gz}{here}.}

\item{gene}{A URL to a tab separated file that maps transcript identifiers
to common gene names. An example file can be found
\href{http://hgdownload.soe.ucsc.edu/goldenPath/hg38/database/kgXref.txt.gz}{here}.}

\item{tx_cols}{A character vector of expected column names for the
known-gene reference file. Required columns: "tx", "chr", "strand",
"cds_start", "cds_end", "exon_start" and "exon_end". All other columns will
be ignored.}

\item{gene_cols}{A character vector of expected column names for the
cross-reference file. Required columns: "tx" and "gene". All other columns
will be ignored.}

\item{shift_start}{Number of bases to shift the start positions. Defaults
to 1 as this is necessary for compatibility with \link[Biostrings:getSeq]{Biostrings::getSeq} which
includes the start position in the returned sequence and begins counting bases
at 1.}

\item{shift_end}{Number of bases to shift the end positions. Defaults to 0.}
}
\value{
A data.frame with the following columns where each row represents
a single exon:
\itemize{
\item \strong{COLUMN-NAME} \emph{\code{DATA-TYPE}} DESCRIPTION
\item \strong{tx}      \emph{\code{chr}} Transcript symbol
\item \strong{gene}    \emph{\code{chr}} Gene symbol
\item \strong{exon}    \emph{\code{int}} Exon rank in gene (lowest contains ATG, highest contains native Stop)
\item \strong{chr}     \emph{\code{chr}} Chromosome
\item \strong{strand}  \emph{\code{chr}} Strand (+/-)
\item \strong{start}   \emph{\code{int}} CDS coordinate start (always <= \strong{end})
\item \strong{end}     \emph{\code{int}} CDS coordinate end (always >= \strong{start})
}
}
\description{
\code{CDS} takes transcript annotation tables in \href{https://genome.ucsc.edu/cgi-bin/hgTables}{UCSC format}
and reshapes them to have coordinates for each exon represented on a single
row, rather than collapsed into a comma separated string in a single cell.
}
\details{
The output of \code{CDS} should meet the following standards (1) each row should
represent the coordinates of a single exon, (2) exons should be numbered in
order with reference to the transcript's strand (e.g. the first exon should
include the start codon). The absolute numbering is unimportant so long as they
are numbered in the correct order. (3) The first and last exon coordinates should
begin with the start codon and end with the stop codon.

To save the trouble of looking up URLs, pre-defined CDS builders
are provided. They are named \code{CDS_<Species>_<data-source>_<genome-assembly-ID>()}.
}
