% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate-PAM.R
\name{locate_PAM}
\alias{locate_PAM}
\title{Locate PAM near a genomic coordinate}
\usage{
locate_PAM(codons, genome, spacing = c(12, 16), PAM = c(sgNGG = ".GG", sgNGA
  = ".GA", sgNGCG = ".GCG", sgNGAG = ".GAG", sgNNGRRT = "..G[AG][AG]T", sgNNNRRT
  = "...[AG][AG]T"), PAM_widths = c(3, 3, 4, 4, 6, 6), flanking = 150)
}
\arguments{
\item{codons}{A dataframe resulting from \link{locate_codons}.}

\item{genome}{A \link[BSgenome:BSgenome]{BSgenome} sequence database,
or a \link[=readDNAStringSet]{Biostrings}. Used to extract the genomic sequence context
for each genomic coordinate in \code{codons}.}

\item{spacing}{A vector of two integers. The range of PAM spacing. These
numbers correspond to the minimum and maximum allowable nucleotides between
the targeted base and the PAM. Defaults to \code{c(12, 16)}}

\item{PAM}{A named vector of PAM motifs to be considered. Names correspond to
resulting column names in the returned dataframe. Values correspond to regular
expressions used to identify the PAM. Default PAM patterns include \code{.GG}, \code{.GA}
\code{.GCG}, \code{.GAG}, \code{..G[AG][AG]T}, and \code{...[AG][AG]T}.}

\item{PAM_widths}{A vector with one integer for every \code{PAM}. These numbers correspond
to the width of the PAM.}

\item{flanking}{An integer specifying how much flanking genomic context to return
in the resulting dataframe. Defaults to 150.}
}
\description{
Uses the results of \link{locate_codons} to determine whether there is an available
PAM with desired spacing from the genomic coordinate of the targeted base.
}
