% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-isoforms.R
\name{plot_spliced_isoforms}
\alias{plot_spliced_isoforms}
\title{Plot spliced isoforms}
\usage{
plot_spliced_isoforms(gene, coords, colors = NULL, ...)
}
\arguments{
\item{gene}{A single gene name}

\item{coords}{A dataframe of \link[=CDS]{CDS coordinates} (will be filtered for \code{gene})}

\item{colors}{A character vector of colors to be applied to each track}

\item{...}{Tracks. Each track must be a dataframe with two columns - \code{gene} and \code{genome_coord}.}
}
\description{
Plot spliced isoforms
}
